Exception handling - delete

+++ src/main/java/com/valbtecnologia/course/services/UserService.java
@@ -4,10 +4,13 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.valbtecnologia.course.entities.User;
 import com.valbtecnologia.course.repositories.UserRepository;
+import com.valbtecnologia.course.services.exceptions.DatabaseException;
 import com.valbtecnologia.course.services.exceptions.ResourceNotFoundException;
 
 
@@ -32,7 +35,17 @@ public class UserService {
 	}
 	
 	public void delete(Long id) {
-		repository.deleteById(id);
+		try {
+			if (!repository.existsById(id)) {
+		        System.out.println("Erro: Usuário não encontrado!");
+		        throw new ResourceNotFoundException(id);
+		    }
+			repository.deleteById(id);
+		} catch (EmptyResultDataAccessException e) {
+			throw new ResourceNotFoundException(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
 	}
 	
 	public User update(Long id, User obj) {
